refactor: move `Tag` classes under `Version`
These classes are around tagging a specific **version**, and thus should
be in that namespace.

This includes namespacing the command as `version:tag` (instead of
`tag`).
